refactor: cria ConfirmationDialogComponent especializado na modal de confirmação
