+++ src/app/features/list/list.component.ts
@@ -4,36 +4,8 @@ import { Product } from '../../interfaces/product';
 import { CardComponent } from '../../components/card/card.component';
 import { Router, RouterLink } from '@angular/router';
 import { MatButtonModule } from '@angular/material/button';
-import {
-  MatDialog,
-  MatDialogModule,
-  MatDialogRef,
-} from '@angular/material/dialog';
-
-@Component({
-  selector: 'app-comfirmation-dialog',
-  template: `
-    <h2 mat-dialog-title>Deletar Produto</h2>
-    <mat-dialog-content>
-      Deseja mesmo deletar este produto?
-    </mat-dialog-content>
-    <mat-dialog-actions align="end">
-      <button mat-button (click)="onNot()">Não</button>
-      <button mat-button (click)="onYes()" cdkFocusInitial>Sim</button>
-    </mat-dialog-actions>
-  `,
-  standalone: true,
-  imports: [MatButtonModule, MatDialogModule],
-})
-export class ConfirmationDialogComponent {
-  matDialogRef = inject(MatDialogRef);
-  onNot() {
-    this.matDialogRef.close(false);
-  }
-  onYes() {
-    this.matDialogRef.close(true);
-  }
-}
+import { ConfirmationDialogService } from '../../services/confirmation-dialog.service';
+import { filter } from 'rxjs';
 
 @Component({
   selector: 'app-list',
@@ -46,7 +18,7 @@ export class ListComponent {
   products: Product[] = [];
   productsService = inject(ProductsService);
   router = inject(Router);
-  matDailog = inject(MatDialog);
+  confirmationDialogService = inject(ConfirmationDialogService);
 
   ngOnInit() {
     this.productsService.getAll().subscribe((products) => {
@@ -57,17 +29,15 @@ export class ListComponent {
     this.router.navigate(['/edit-product', product.id]);
   }
   onDelete(product: Product) {
-    this.matDailog
-      .open(ConfirmationDialogComponent)
-      .afterClosed()
-      .subscribe((answer: boolean) => {
-        if (answer) {
-          this.productsService.delete(product.id).subscribe(() => {
-            this.productsService.getAll().subscribe((products) => {
-              this.products = products;
-            });
-          });
-        }
+    this.confirmationDialogService
+    .openDialog()
+    .pipe(filter((answer) => answer === true))
+    .subscribe(() => {
+      this.productsService.delete(product.id).subscribe(() => {
+        this.productsService.getAll().subscribe((products) => {
+          this.products = products;
+        });
       });
+    });
   }
 }
