+++ src/app/services/confirmation-dialog.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { ConfirmationDialogService } from './confirmation-dialog.service';
+
+describe('ConfirmationDialogService', () => {
+  let service: ConfirmationDialogService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(ConfirmationDialogService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});
